+++ packages/web-visualization/CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 - Support connecting null values in Area and Line
 - Added label to XAxis and YAxis
 - Added gradient support
+- Improved charts accessibility support
 
 ## 3.4.0-beta.5 (11/4/2025 PST)